requirements.txt
pyautogui
pygetwindow
pillow
numpy
astar_python